Add custom server validation date availability
Added custom validation for when a user wants to book a new appointment. The validation checks if the submitted date is available and doesn't let the user create a booking if it isn't.
PS. Changed some minor things in some entities (the ID of a Booking from and int to Guid).